requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aioprocessing==2.0.1
aiosignal==1.3.1
analytics-python==1.2.9
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
docker-pycreds==0.4.0
emoji==2.12.1
exceptiongroup==1.2.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
gql==3.5.0
graphql-core==3.2.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
janus==1.0.0
lancedb==0.10.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
openai==1.37.0
overrides==7.7.0
packaging==24.1
platformdirs==4.2.2
protobuf==5.27.2
psutil==6.0.0
py==1.11.0
pyarrow==15.0.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pylance==0.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
PyYAML==6.0.1
ratelimiter==1.2.0.post0
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.7.1
sentry-sdk==2.11.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
wandb==0.17.5
weave==0.50.11
Werkzeug==3.0.3
yarl==1.9.4
